+++ src/trace_processor/shell/BUILD.gn
@@ -20,8 +20,6 @@ source_set("shell") {
     "common_flags.h",
     "convert_subcommand.cc",
     "convert_subcommand.h",
-    "ctl_subcommand.cc",
-    "ctl_subcommand.h",
     "export_subcommand.cc",
     "export_subcommand.h",
     "interactive.cc",
