+++ src/trace_processor/trace_processor_shell.cc
@@ -54,7 +54,6 @@
 #include "src/trace_processor/rpc/stdiod.h"
 #include "src/trace_processor/shell/common_flags.h"
 #include "src/trace_processor/shell/convert_subcommand.h"
-#include "src/trace_processor/shell/ctl_subcommand.h"
 #include "src/trace_processor/shell/export_subcommand.h"
 #include "src/trace_processor/shell/interactive_subcommand.h"
 #include "src/trace_processor/shell/metatrace.h"
@@ -161,7 +160,6 @@ Commands:
   export        Export a trace to a database file.
   metrics       Run v1 metrics (deprecated; use 'summarize --metrics-v2').
   convert       Convert trace format.
-  ctl           Manage warm sessions (e.g. ctl kill-server <name>).
 
 Common flags (apply to all commands):
   -h, --help                  Show help (per-command if after a command).
@@ -791,11 +789,10 @@ base::Status TraceProcessorShell::Run(int argc, char** argv) {
     shell::ExportSubcommand export_subcommand;
     shell::MetricsSubcommand metrics_subcommand;
     shell::ConvertSubcommand convert_subcommand;
-    shell::CtlSubcommand ctl_subcommand;
     std::vector<shell::Subcommand*> subcommands = {
         &query_subcommand,     &interactive_subcommand, &server_subcommand,
         &summarize_subcommand, &export_subcommand,      &metrics_subcommand,
-        &convert_subcommand,   &ctl_subcommand,
+        &convert_subcommand,
     };
 
     // Handle "help" pseudo-subcommand: `tp help <command>` or bare `tp help`.
